Formalise code: bits and bobs
- Got rid of unnecessary reinterpret casts,
- Improved const-correctness,
- Declared some variables out of for-loops (instead of using glm::vec3 constructor repeatedly for instance),
- Used vector references and reserved memory beforehand in some places.

+++ OpenGLProject/OpenGLProject/VertexArray.cpp
@@ -45,7 +45,7 @@ void VertexArray::AddBuffer(const VertexBuffer& vbo, const VertexBufferLayout& l
 		glEnableVertexAttribArray(attributeLayout.location);
 
 		// Store in the VAO how we want OpenGL to interpret the VBO data relative to the attribute index (layout)
-		glVertexAttribPointer(attributeLayout.location, attributeLayout.count, attributeLayout.type, attributeLayout.normalized, layout.GetStride(), reinterpret_cast<const void*>(static_cast<const size_t>(offset)));
+		glVertexAttribPointer(attributeLayout.location, attributeLayout.count, attributeLayout.type, attributeLayout.normalized, layout.GetStride(), reinterpret_cast<const void*>(offset));
 
 		offset += attributeLayout.count * sizeof(attributeLayout.type);
 	}
@@ -65,7 +65,7 @@ void VertexArray::AddInstancedBuffer(const VertexBuffer& vbo, const VertexBuffer
 		glEnableVertexAttribArray(attributeLayout.location);
 
 		// Store in the currently bound VBO how we want OpenGL to interpret the data corresponding to the attribute index (layout)
-		glVertexAttribPointer(attributeLayout.location, attributeLayout.count, attributeLayout.type, attributeLayout.normalized, sizeof(glm::mat4), reinterpret_cast<const void*>(static_cast<size_t>(offset)));
+		glVertexAttribPointer(attributeLayout.location, attributeLayout.count, attributeLayout.type, attributeLayout.normalized, sizeof(glm::mat4), reinterpret_cast<const void*>(offset));
 
 		// Set the rate at which the attribute index advance when rendering multiple instances of primitives in a single draw call
 		glVertexAttribDivisor(attributeLayout.location, 1);
